using custom hook for the flow
